Support fro displaying the filament width on an 20x4 LCD
Changes to support displaying the real-time filament width and the
volume factor on a 20x4 LCD.  The data is displayed on the 4th line.
First the status message is displayed for 5 seconds, and then the
filament data is displayed.  The status message can be seen by
re-selecting the info screen in the menu.  Also includes the change to
use SET_INPUT or SET_OUTPUT instead of pinMode to be compatible with
teensy.